DEV: Move template-tag to bottom of class definition (#41)
This seems to be the more common convention in the wider ember community. Arguably it makes more sense because you define the fields/getters/functions first, and then make use of them in the `<template>`.

Unfortunately the auto-fix for this rule doesn't work properly for template-tag, so we'll have to fix this up manually in our gjs files.