Introduce `discourse/no-implicit-this` with autofix (#84)
This is copy/pasted from upstream's `no-implicit-this` rule, with the addition of autofix functionality. Upstream likely decided against autofix because it's possible for this rule to raise false-positives, which then require helpers to be added to the 'allow' list. However, in our case, the allow list is preconfigured with core helpers, and it's exceptionally rare for plugins/themes to introduce their own argument-less helpers.